myproject/requirements.txt
pandas
python-dotenv
apscheduler
openpyxl
pytz
 